WIP cached scoring
